+++ drizzle/0001_nappy_valeria_richards.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "users" ADD COLUMN "lud16" varchar(255);